Inline buffer sources in buffer.c
